JUCE対応 : lsp::synth::Voice を juce::SynthesizerVoiceベースへ置換
