--- libsynthpp/src/lsp/synth/channel_params_melody.cpp
@@ -1,235 +1,0 @@
-﻿/**
-	libsynth++
-
-	Copyright(c) 2018 my04337
-
-	This software is released under the MIT License.
-	https://opensource.org/license/mit/
-*/
-
-#include <lsp/synth/channel_params.hpp>
-#include <lsp/synth/synth.hpp>
-#include <lsp/synth/instruments.hpp>
-
-using namespace lsp::synth;
-
-
-static const std::unordered_map<
-	int, std::tuple<
-	float, // v: volume(adjuster)
-	float, // a: sec
-	float, // h: sec
-	float, // d: sec
-	float, // s: level
-	float, // f: Linear : level/sec, Exp : dBFS/sec
-	float // r: sec
-	>
-> sMelodyParams = {
-	{0,   { 1.00f, 0.02f, 0.00f,  3.00f, 0.00f, 0.00f, 1.00f }},
-	{1,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{2,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{3,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{4,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{5,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{6,   { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 1.00f }},
-	{7,   { 1.00f, 0.02f, 0.00f,  6.00f, 0.00f, 0.00f, 0.10f }},
-	{8,   { 1.00f, 0.02f, 0.00f,  4.00f, 0.00f, 0.00f, 1.50f }},
-	{9,   { 1.00f, 0.02f, 0.00f,  4.00f, 0.00f, 0.00f, 2.00f }},
-	{10,  { 1.00f, 0.02f, 0.00f,  4.00f, 0.00f, 0.00f, 2.00f }},
-	{11,  { 1.00f, 0.02f, 0.00f,  7.00f, 0.00f, 0.00f, 1.00f }},
-	{12,  { 1.00f, 0.02f, 0.00f,  3.00f, 0.00f, 0.00f, 2.00f }},
-	{13,  { 1.00f, 0.02f, 0.00f,  3.00f, 0.00f, 0.00f, 2.00f }},
-	{14,  { 1.00f, 0.02f, 0.00f,  7.00f, 0.00f, 0.00f, 5.00f }},
-	{15,  { 1.00f, 0.02f, 0.00f,  3.00f, 0.00f, 0.00f, 2.00f }},
-	{16,  { 1.00f, 0.02f, 0.00f,  0.50f, 0.70f, 0.00f, 0.10f }},
-	{17,  { 1.00f, 0.02f, 0.00f,  0.50f, 0.70f, 0.00f, 0.10f }},
-	{18,  { 1.00f, 0.02f, 0.00f,  0.50f, 0.70f, 0.00f, 0.10f }},
-	{19,  { 1.00f, 0.02f, 0.00f,  0.50f, 0.70f, 0.00f, 1.00f }},
-	{20,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.50f, 0.00f, 1.00f }},
-	{21,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{22,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{23,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{24,  { 1.00f, 0.02f, 0.00f,  7.00f, 0.00f, 0.00f, 0.10f }},
-	{25,  { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 0.20f }},
-	{26,  { 1.00f, 0.02f, 0.00f,  6.00f, 0.00f, 0.00f, 0.20f }},
-	{27,  { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 0.20f }},
-	{28,  { 1.00f, 0.02f, 0.00f,  5.00f, 0.00f, 0.00f, 0.20f }},
-	{29,  { 1.00f, 0.02f, 0.00f,  7.00f, 0.08f, 0.00f, 0.10f }},
-	{30,  { 1.00f, 0.02f, 0.20f,  7.00f, 0.08f, 0.00f, 0.10f }},
-	{31,  { 1.00f, 0.02f, 0.20f, 10.00f, 0.00f, 0.00f, 0.10f }},
-	{32,  { 1.00f, 0.02f, 0.00f,  5.00f, 0.08f, 0.00f, 0.10f }},
-	{33,  { 1.00f, 0.02f, 0.00f,  5.00f, 0.03f, 0.00f, 0.10f }},
-	{34,  { 1.00f, 0.02f, 0.00f,  4.00f, 0.06f, 0.00f, 0.10f }},
-	{35,  { 1.00f, 0.02f, 0.00f, 10.00f, 0.00f, 0.00f, 0.10f }},
-	{36,  { 1.00f, 0.02f, 0.00f,  3.00f, 0.01f, 0.00f, 0.10f }},
-	{37,  { 1.00f, 0.02f, 0.00f,  3.00f, 0.01f, 0.00f, 0.10f }},
-	{38,  { 1.00f, 0.02f, 0.00f,  1.50f, 0.07f, 0.00f, 0.10f }},
-	{39,  { 1.00f, 0.02f, 0.20f,  5.00f, 0.13f, 0.00f, 0.10f }},
-	{40,  { 1.00f, 0.03f, 0.20f,  5.00f, 0.66f, 0.00f, 0.20f }},
-	{41,  { 1.00f, 0.03f, 0.35f,  5.00f, 0.66f, 0.00f, 0.20f }},
-	{42,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.20f }},
-	{43,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.20f }},
-	{44,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.90f, 0.00f, 1.00f }},
-	{45,  { 1.00f, 0.03f, 0.00f,  0.50f, 0.00f, 0.00f, 0.50f }},
-	{46,  { 1.00f, 0.02f, 0.00f,  4.00f, 0.00f, 0.00f, 3.00f }},
-	{47,  { 3.50f, 0.02f, 0.00f,  0.30f, 0.00f, 0.00f, 0.30f }},
-	{48,  { 1.00f, 0.02f, 0.00f,  0.30f, 0.80f, 0.00f, 0.70f }},
-	{49,  { 1.00f, 0.03f, 0.00f,  0.30f, 0.80f, 0.00f, 1.20f }},
-	{50,  { 1.00f, 0.03f, 0.00f,  0.30f, 0.80f, 0.00f, 0.70f }},
-	{51,  { 1.00f, 0.03f, 0.00f,  0.30f, 0.80f, 0.00f, 1.20f }},
-	{52,  { 1.00f, 0.03f, 0.00f,  0.30f, 0.80f, 0.00f, 0.70f }},
-	{53,  { 1.00f, 0.03f, 0.00f,  0.30f, 0.80f, 0.00f, 0.70f }},
-	{54,  { 1.00f, 0.03f, 0.00f,  0.50f, 0.80f, 0.00f, 0.70f }},
-	{55,  { 2.00f, 0.02f, 0.10f,  0.60f, 0.00f, 0.00f, 0.40f }},
-	{56,  { 1.00f, 0.03f, 0.00f,  0.50f, 0.60f, 0.00f, 0.10f }},
-	{57,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.60f, 0.00f, 0.10f }},
-	{58,  { 1.00f, 0.03f, 0.00f,  2.00f, 0.60f, 0.00f, 0.10f }},
-	{59,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.60f, 0.00f, 0.10f }},
-	{60,  { 1.00f, 0.03f, 0.00f,  2.00f, 0.60f, 0.00f, 0.10f }},
-	{61,  { 1.00f, 0.03f, 0.00f,  1.50f, 0.50f, 0.00f, 0.10f }},
-	{62,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.50f, 0.00f, 0.10f }},
-	{63,  { 1.00f, 0.03f, 0.00f,  2.00f, 0.50f, 0.00f, 0.10f }},
-	{64,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{65,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{66,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{67,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{68,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{69,  { 1.00f, 0.03f, 0.00f,  0.50f, 0.80f, 0.00f, 0.10f }},
-	{70,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{71,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{72,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{73,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{74,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{75,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{76,  { 1.00f, 0.10f, 0.00f,  0.50f, 0.80f, 0.00f, 0.10f }},
-	{77,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.10f }},
-	{78,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.30f }},
-	{79,  { 1.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.10f }},
-	{80,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{81,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.80f, 0.00f, 0.10f }},
-	{82,  { 1.00f, 0.02f, 0.00f,  1.00f, 0.60f, 0.00f, 0.10f }},
-	{83,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.25f, 0.00f, 0.10f }},
-	{84,  { 1.00f, 0.03f, 0.03f,  4.00f, 0.25f, 0.00f, 0.10f }},
-	{85,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.56f, 0.00f, 0.20f }},
-	{86,  { 1.00f, 0.03f, 0.15f,  4.00f, 0.31f, 0.00f, 0.30f }},
-	{87,  { 1.00f, 0.03f, 0.00f,  4.50f, 0.31f, 0.00f, 0.10f }},
-	{88,  { 1.00f, 0.03f, 0.00f,  3.00f, 0.56f, 0.00f, 3.00f }},
-	{89,  { 1.00f, 0.03f, 0.00f,  3.00f, 0.90f, 0.00f, 1.00f }},
-	{90,  { 1.00f, 0.03f, 0.00f,  3.00f, 0.40f, 0.00f, 0.80f }},
-	{91,  { 1.00f, 0.03f, 0.00f,  3.00f, 0.90f, 0.00f, 3.00f }},
-	{92,  { 1.00f, 0.03f, 0.00f,  3.00f, 0.90f, 0.00f, 2.50f }},
-	{93,  { 1.00f, 0.30f, 0.00f,  8.00f, 0.00f, 0.00f, 2.00f }},
-	{94,  { 1.00f, 0.03f, 0.15f,  1.00f, 0.90f, 0.00f, 1.00f }},
-	{95,  { 1.00f, 0.30f, 0.00f,  5.00f, 0.50f, 0.00f, 2.00f }},
-	{96,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.50f, 0.00f, 2.00f }},
-	{97,  { 1.00f, 0.03f, 0.10f,  3.00f, 0.70f, 0.00f, 3.00f }},
-	{98,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.00f, 0.00f, 3.00f }},
-	{99,  { 1.00f, 0.03f, 0.00f,  4.00f, 0.10f, 0.00f, 2.00f }},
-	{100, { 1.00f, 0.03f, 0.10f,  5.00f, 0.00f, 0.00f, 3.00f }},
-	{101, { 1.00f, 1.00f, 0.00f,  1.00f, 0.90f, 0.00f, 2.00f }},
-	{102, { 1.00f, 0.03f, 0.00f,  1.00f, 0.90f, 0.00f, 2.00f }},
-	{103, { 1.00f, 0.03f, 0.00f,  6.00f, 0.04f, 0.00f, 2.00f }},
-	{104, { 1.00f, 0.03f, 0.03f,  9.00f, 0.00f, 0.00f, 1.00f }},
-	{105, { 1.00f, 0.03f, 0.00f,  6.00f, 0.00f, 0.00f, 1.00f }},
-	{106, { 1.00f, 0.03f, 0.00f,  3.00f, 0.00f, 0.00f, 1.00f }},
-	{107, { 1.00f, 0.03f, 0.00f,  3.00f, 0.00f, 0.00f, 2.00f }},
-	{108, { 1.00f, 0.03f, 0.00f,  1.50f, 0.00f, 0.00f, 1.00f }},
-	{109, { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{110, { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{111, { 1.00f, 0.03f, 0.00f,  1.00f, 0.70f, 0.00f, 0.10f }},
-	{112, { 1.00f, 0.03f, 0.00f,  3.00f, 0.00f, 0.00f, 1.00f }},
-	{113, { 1.00f, 0.03f, 0.00f,  0.40f, 0.00f, 0.00f, 0.30f }},
-	{114, { 1.00f, 0.03f, 0.00f,  3.00f, 0.00f, 0.00f, 2.00f }},
-	{115, { 1.00f, 0.03f, 0.00f,  0.50f, 0.00f, 0.00f, 0.50f }},
-	{116, { 1.00f, 0.03f, 0.00f,  3.00f, 0.00f, 0.00f, 3.00f }},
-	{117, { 1.00f, 0.03f, 0.00f,  0.80f, 0.00f, 0.00f, 0.80f }},
-	{118, { 1.00f, 0.03f, 0.00f,  0.80f, 0.00f, 0.00f, 0.80f }},
-	{119, { 1.00f, 2.00f, 0.00f,  0.10f, 0.00f, 0.00f, 0.10f }},
-	{120, { 1.00f, 0.03f, 0.00f,  0.50f, 0.00f, 0.00f, 0.50f }},
-	{121, { 1.00f, 0.03f, 0.00f,  0.50f, 0.00f, 0.00f, 0.40f }},
-	{122, { 2.00f, 0.50f, 1.50f,  3.00f, 0.00f, 0.00f, 1.50f }},
-	{123, { 2.00f, 0.50f, 0.80f,  4.00f, 0.00f, 0.00f, 0.50f }},
-	{124, { 2.00f, 0.03f, 0.00f,  1.00f, 0.80f, 0.00f, 0.10f }},
-	{125, { 2.00f, 3.50f, 0.00f,  1.00f, 0.80f, 0.00f, 0.70f }},
-	{126, { 2.00f, 2.00f, 0.00f,  1.00f, 0.80f, 0.00f, 0.80f }},
-	{127, { 2.50f, 0.03f, 0.00f,  1.50f, 0.00f, 0.00f, 1.50f }},
-};
-
-std::unique_ptr<LuathVoice> ChannelParams::createMelodyVoice(int noteNo, float vel)
-{
-	float v = 1.f; // volume(adjuster)
-	float a = 0.f; // sec
-	float h = 0.f; // sec
-	float d = 0.f; // sec
-	float s = 1.f; // level
-	float f = 0.f; // Linear : level/sec, Exp : dBFS/sec
-	float r = 0.f; // sec
-
-	// 主要パラメータのロード
-	if(auto found = sMelodyParams.find(mProgId); found != sMelodyParams.end()) {
-		std::tie(v, a, h, d, s, f, r) = found->second;
-	}
-
-	// リズム系の楽器は若干パラメータを補正 ※ドラムほど強くパラメータは変更しない
-	bool isDrumLikeInstrument = false;
-	float adjustedNoteNo = static_cast<float>(noteNo);
-	switch(mProgId)
-	{
-	case 47:
-	case 121:
-	case 122:
-	case 125:
-	case 126:
-	case 127:
-		isDrumLikeInstrument = true;
-		adjustedNoteNo -= 12; // 1オクターブ下げる
-		break;
-	}
-
-	// 音色の調整
-	dsp::WaveTableGenerator<float> wg;
-	if(isDrumLikeInstrument) {
-		wg = Instruments::createDrumNoiseGenerator();
-	} else{
-		// 素の正弦波は結構どぎつい倍音を持つため、折り返しノイズ等を避けるために大まかなノート番号によって倍音成分を変化させる
-		// ※正弦波の倍音は1～50まで用意してあるため、良い感じにマッピングする
-		static constexpr float min_note_no = 0;  // C-1 : 8.2[Hz]
-		static constexpr float max_note_no = 108; // C8 : 4186.0[Hz]
-		static constexpr int min_dim = 5; 
-		static constexpr int max_dim = 30;
-
-		auto waveTableId = std::clamp(
-			static_cast<int>(min_dim + (1.f - (adjustedNoteNo - min_note_no) / (max_note_no - min_note_no)) * (max_dim - min_dim)),
-			min_dim,
-			max_dim
-		);
-		wg = Instruments::createSquareGenerator(waveTableId);
-	}
-
-
-	if(mSystemType.isGS() || mSystemType.isXG()) {
-		a *= powf(10.0f, (ccAttackTime / 128.f - 0.5f) * 4.556f);
-		d *= powf(10.0f, (ccDecayTime / 128.f - 0.5f) * 4.556f);
-		r *= powf(10.0f, (ccReleaseTime / 128.f - 0.5f) * 4.556f);
-	}
-
-	// MEMO 人間の聴覚ではボリュームは対数的な特性を持つため、ベロシティを指数的に補正する
-	// TODO sustain_levelで除算しているのは旧LibSynth++からの移植コード。 補正が不要になったら削除すること
-	float volume = powf(10.f, -20.f * (1.f - vel) / 20.f) * v / ((s > 0.8f && s != 0.f) ? s : 0.8f);
-	float cutoffLevel = 0.01f;
-	static const dsp::EnvelopeGenerator<float>::Curve curveExp3(3.0f);
-
-	float overtuneGain = 0.f; // dB
-	if(mSystemType.isGS() || mSystemType.isXG()) {
-		overtuneGain = getInt7NRPN(1, 33).value_or(0) / 128.f * 5.f;
-	}
-
-	auto voice = std::make_unique<LuathVoice>(wg, adjustedNoteNo, mCalculatedPitchBend, volume, ccPedal);
-	voice->setCurrentPlaybackSampleRate(mSynth.sampleFreq());
-	voice->setResonance(2.f, overtuneGain);
-
-	auto& eg = voice->envolopeGenerator();
-	eg.setMelodyEnvelope(mSynth.sampleFreq(), curveExp3,a, h, d, s, f, r, cutoffLevel);
-	eg.noteOn();
-	return voice;
-}
