+++ libsynthpp/src/lsp/synth/voice.hpp
@@ -18,23 +18,8 @@
 namespace lsp::synth
 {
 class LuathSynth;
+struct ChannelState;
 
-// Luath用サウンド 
-// MEMO juce::SynthesizerVoiceは何らかのjuce::SynthesizerSoundに属している必要があるため、最低限の物を用意した。
-class LuathSound final
-	: public juce::SynthesiserSound
-{
-public:
-	LuathSound(LuathSynth& synth) : mSynth(synth) {}
-
-	bool appliesToNote(int midiNoteNumber)override { return false; }
-	bool appliesToChannel(int midiChannel)override { return false; }
-
-	LuathSynth& synth()const noexcept { return mSynth; }
-
-private:
-	LuathSynth& mSynth;
-};
 
 // ボイス識別番号
 struct _voice_id_tag {};
@@ -46,6 +31,8 @@ class LuathVoice final
 	: public juce::SynthesiserVoice
 	, non_copy_move
 {
+	using SUPER = juce::SynthesiserVoice;
+
 public:
 	using EnvelopeGenerator = dsp::EnvelopeGenerator<float>;
 	using EnvelopeState = dsp::EnvelopeState;
@@ -53,64 +40,52 @@ public:
 	using WaveTableGenerator = dsp::WaveTableGenerator<float>;
 
 	struct Digest {
+		int ch = 1; // チャネル [1～16]
 		float freq = 0; // 基本周波数
 		float envelope = 0; // エンベロープジェネレータ出力
 		dsp::EnvelopeState state = dsp::EnvelopeState::Free; // エンベロープジェネレータ ステート
 	};
 
 public:
-	LuathVoice(const WaveTableGenerator& wg, float noteNo, float pitchBend, float volume, bool hold);
+	LuathVoice(LuathSynth& synth);
 	virtual ~LuathVoice();
 
 	// --- implmentation of uce::SynthesiserVoice --- 
 	bool canPlaySound(juce::SynthesiserSound*)override;
-	void startNote(int midiNoteNumber, float velocity, juce::SynthesiserSound* sound, int currentPitchWheelPosition) {} // TODO
-	void stopNote(float velocity, bool allowTailOff) {} // TODO
-	void pitchWheelMoved(int newPitchWheelValue) {} // TODO
+	bool isVoiceActive()const override;
+	void startNote(int noteNo, float velocity, juce::SynthesiserSound* sound, int currentPitchWheelPosition);
+	void stopNote(float velocity, bool allowTailOff);
+	void pitchWheelMoved(int newPitchWheelValue);
 	void controllerMoved(int controllerNumber, int newControllerValue) {} // TODO
-	void renderNextBlock(juce::AudioBuffer< float >& outputBuffer, int startSample, int numSamples) {} // TODO
+	void renderNextBlock(juce::AudioBuffer<float>& outputBuffer, int startSample, int numSamples);
 	// ---
-
-	float update()
-	{
-		auto v = mWG.update(sampleFreq(), mCalculatedFreq);
-		v = mCutOffFilter.update(v);
-		v = mResonanceFilter.update(v);
-		v *= mEG.update();
-		v *= mVolume;
-		return v;
-	}
 
 	Digest digest()const noexcept;
 	float sampleFreq()const noexcept { return static_cast<float>(getSampleRate()); }
 
-	float noteNo()const noexcept;
-	void noteOff(bool allowTailOff = true)noexcept;
-
-	std::optional<float> pan()const noexcept;
-	void setPan(float pan)noexcept;
-
-	void setPitchBend(float pitchBend)noexcept;
 	EnvelopeGenerator& envolopeGenerator() noexcept;
 
 	void setCutOff(float freqRate, float cutOffGain);
 	void setResonance(float freqRate, float overtoneGain);
 
 
-protected:
-	void updateFreq()noexcept;
+private:
+	const ChannelState& getChannelState()const noexcept;
+	void prepareMelodyNote(int noteNo, float velocity, int currentPitchWheelPosition);
+	void prepareDrumNote(int noteNo, float velocity, int currentPitchWheelPosition);
 
-protected:
+private:
+	LuathSynth& mSynth;
 	EnvelopeGenerator mEG;
 	BiquadraticFilter mCutOffFilter;
 	BiquadraticFilter mResonanceFilter;
 	WaveTableGenerator mWG;
-	float mNoteNo;
-	bool mPendingNoteOff = false;
-	float mPitchBend;
-	float mCalculatedFreq = 0;
-	float mVolume;
-	std::optional<float> mPan; // ドラムなど、ボイス毎にパンが指定される場合のヒント
+	float mWaveTableNoteNo = 0; // [NoteNo]
+	float mPitchBend = 0;       // [NoteNo]
+	float mResolvedFreq = 0;    // [Hz]
+	float mVolume = 0;
+	float mChannelPan = 0.5f; // チャネル全体のパン
+	std::optional<float> mPerVoicePan; // ドラムなど、ボイス毎にパンが指定される場合のヒント
 };
 
 }
